Add a test for a NI postcode on a NI licence
The intent of this test is to ensure that Frontend is able to identify a
Northern Ireland local authority correctly in the MapIt response. Many of our
location-based formats don't support Northern Ireland for various reasons, but
licences should. To enable that we added the GSS codes for the NI LAs as SNAC
codes in MapIt [1], Frontend [2] and Licensify [3]. This check will alert us
if we forget to add them in again when we update our MapIt database in the
future.

[1]: https://github.com/alphagov/mapit/pull/13
[2]: https://github.com/alphagov/frontend/pull/918
[3]: This was done manually in the Licensify admin interface.